move pagination css into folder
