+++ src/global/components/Pagination/css/Pagination.css
