+++ src/global/components/Pagination/css/PaginationTotal.css
